Implement subproccess creation and piping behaviour
We now start a subprocess to handle the majority of the parsing. We set up
pipes to and from the subprocess so we can send it input and get its output.
If the subproccess ends (or closes it's ends of the pipe), we restart it with
new pipes.

In the event that we have stuff to send to it but do not yet have a child
proccess, we store the output into a cbuffer which then gets emptied when we
get the chance.