Start a bit of rearranging. Touch all the things.
bmap and cbuffer have been moved into a lib folder. The Makefile now build a
static library for use in the other programs.

The aim is to make conbot and jbotc separate as well, but work is being done
in that direction.

The defines now have some of the conbot defines as well. Things have been made
into actual #defines to prevent errors.