Switch from matching based on strings to regex
Another macro has been added, and one changed for building the functions
array. In order to initialize all the regex_t, a function must be called to
loop over the functions and try to compile them all.

The way functions use arguments will change, as it was a ugly mix of the old
method (strtok on NULL which happened to work) and trying to do things with
fa->args which was not properly setup.