Implement all the methods!
Removed the ensure directory function as we no longer bother with them.

The _dump and _read methods have been retooled to dump/read from a single
file. This means that we now use the methods on bmap which take a FILE *
instead of a char *.

Also, errors are much more catastrophic. To handle this, we have a bunch of
TODO's ;) Seriously, this needs a ton of work.