Initial overhaul of testing framework
The test files are now generic and have values substituted in from the
defines.h file. Yay! other people can test it.

A predefined random seed is used, so we can test the functions which have
"random" output, yay!

Most (if not all) of these tests still need updates due to the new
regex/markov dumping messages as well.